+++ tests/test_checker/test_noqa.py
@@ -171,6 +171,8 @@ SHOULD_BE_RAISED = types.MappingProxyType({
     'WPS358': 1,
     'WPS359': 1,
     'WPS360': 1,
+    'WPS361': 1,
+    'WPS362': 2,
 
     'WPS400': 0,  # defined in ignored violations.
     'WPS401': 0,  # logically unacceptable.
@@ -287,6 +289,7 @@ SHOULD_BE_RAISED = types.MappingProxyType({
     'WPS612': 1,
     'WPS613': 1,
     'WPS614': 1,
+    'WPS615': 2,
 })
 
 # Violations which may be tweaked by `i_control_code` option: