+++ tests/test_visitors/test_ast/test_blocks/test_control_variables.py
@@ -43,6 +43,14 @@ def wrapper():
     print(wrapper)
 """
 
+correct_for_multi_loops = """
+def wrapper():
+    for my_var in range(3):
+        my_var = my_var + 3
+    for my_var in range(4):
+        my_var = my_var + 4
+"""
+
 correct_for_comprehension1 = """
 def context():
     nodes = [
@@ -193,6 +201,14 @@ def wrapper():
     print(second)
 """
 
+wrong_block_variable_reuse1 = """
+def wrapper():
+    for my_var in range(3):
+        my_var = my_var + 3
+    with open() as my_var:
+        my_var = my_var + 4
+"""
+
 
 @pytest.mark.parametrize('code', [
     wrong_for_loop1,
@@ -201,6 +217,7 @@ def wrapper():
     wrong_with1,
     wrong_with2,
     wrong_with3,
+    wrong_block_variable_reuse1,
 ])
 def test_control_variable_used_after_block(
     assert_errors,
@@ -224,6 +241,7 @@ def test_control_variable_used_after_block(
     correct_for_loop3,
     correct_for_loop4,
     correct_for_loop5,
+    correct_for_multi_loops,
     correct_for_comprehension1,
     correct_for_comprehension2,
     correct_for_comprehension3,