+++ wemake_python_styleguide/logic/tree/classes.py
@@ -1,14 +1,22 @@
 import ast
-from typing import List, Optional, Tuple
+from typing import Iterable, List, Optional, Tuple
 
-from wemake_python_styleguide.compat.aliases import AssignNodes
+from typing_extensions import Final
+
+from wemake_python_styleguide.compat.aliases import AssignNodes, FunctionNodes
 from wemake_python_styleguide.constants import ALLOWED_BUILTIN_CLASSES
 from wemake_python_styleguide.logic import nodes
 from wemake_python_styleguide.logic.naming.builtins import is_builtin_name
-from wemake_python_styleguide.types import AnyAssign
+from wemake_python_styleguide.types import AnyAssign, AnyFunctionDef
 
 #: Type alias for the attributes we return from class inspection.
 _AllAttributes = Tuple[List[AnyAssign], List[ast.Attribute]]
+
+#: Prefixes that usually define getters and setters.
+_GetterSetterPrefixes: Final = ('get_', 'set_')
+
+#: Fixes length of a getter/setter.
+GETTER_LENGTH: Final = 4
 
 
 def is_forbidden_super_class(class_name: Optional[str]) -> bool:
@@ -41,7 +49,10 @@ def is_forbidden_super_class(class_name: Optional[str]) -> bool:
     return is_builtin_name(class_name)
 
 
-def get_attributes(node: ast.ClassDef) -> _AllAttributes:
+def get_attributes(
+    node: ast.ClassDef,
+    include_annotated: bool,
+) -> _AllAttributes:
     """
     Helper to get all attributes from class nod definitions.
 
@@ -52,6 +63,7 @@ def get_attributes(node: ast.ClassDef) -> _AllAttributes:
 
     Args:
         node: class node definition.
+        include_annotated: whether or not to include AnnAssign attributes.
 
     Returns:
         A tuple of lists for both class and instance level variables.
@@ -66,7 +78,10 @@ def get_attributes(node: ast.ClassDef) -> _AllAttributes:
             instance_attributes.append(instance_attr)
             continue
 
-        class_attr = _get_class_attribute(node, subnode)
+        if include_annotated:
+            class_attr = _get_annotated_class_attribute(node, subnode)
+        else:
+            class_attr = _get_class_attribute(node, subnode)
         if class_attr is not None:
             class_attributes.append(class_attr)
 
@@ -83,10 +98,33 @@ def _get_instance_attribute(node: ast.AST) -> Optional[ast.Attribute]:
 
 
 def _get_class_attribute(
-    node: ast.ClassDef, subnode: ast.AST,
+    node: ast.ClassDef,
+    subnode: ast.AST,
 ) -> Optional[AnyAssign]:
     return subnode if (
-        isinstance(subnode, AssignNodes) and
         nodes.get_context(subnode) == node and
-        getattr(subnode, 'value', None)
+        getattr(subnode, 'value', None) and
+        isinstance(subnode, AssignNodes)
     ) else None
+
+
+def _get_annotated_class_attribute(
+    node: ast.ClassDef,
+    subnode: ast.AST,
+) -> Optional[AnyAssign]:
+    return subnode if (
+        nodes.get_context(subnode) == node and
+        (
+            getattr(subnode, 'value', None) and
+            isinstance(subnode, AssignNodes)
+        ) or isinstance(subnode, ast.AnnAssign)
+    ) else None
+
+
+def find_getters_and_setters(node: ast.ClassDef) -> Iterable[AnyFunctionDef]:
+    """Returns nodes of all getter or setter methods."""
+    for sub in ast.walk(node):
+        is_correct_context = nodes.get_context(sub) == node
+        if isinstance(sub, FunctionNodes) and is_correct_context:
+            if sub.name[:GETTER_LENGTH] in _GetterSetterPrefixes:
+                yield sub