+++ wemake_python_styleguide/violations/consistency.py
@@ -84,6 +84,8 @@ Summary
    FloatZeroViolation
    UnpackingIterableToListViolation
    RawStringNotNeededViolation
+   InconsistentComprehensionViolation
+   AssignToSliceViolation
 
 Consistency checks
 ------------------
@@ -149,6 +151,8 @@ Consistency checks
 .. autoclass:: FloatZeroViolation
 .. autoclass:: UnpackingIterableToListViolation
 .. autoclass:: RawStringNotNeededViolation
+.. autoclass:: InconsistentComprehensionViolation
+.. autoclass:: AssignToSliceViolation
 
 """
 
@@ -1007,7 +1011,12 @@ class ModuloStringFormatViolation(ASTViolation):
         'some string', 'your name: {0}', 'data: {data}'
 
         # Wrong:
-       'my name is: %s', 'data: %(data)d'
+        'my name is: %s', 'data: %(data)d'
+
+    It might be a good idea to disable this rule
+    and switch to ``flake8-pep3101`` in case your project
+    has a lot of false-positives due
+    to some specific string chars that uses ``%`` a lot.
 
     See also:
         https://github.com/gforcada/flake8-pep3101
@@ -2257,3 +2266,80 @@ class RawStringNotNeededViolation(TokenizeViolation):
 
     error_template = 'Found an unnecessary use of a raw string: {0}'
     code = 360
+
+
+@final
+class InconsistentComprehensionViolation(TokenizeViolation):
+    """
+    Forbids inconsistent newlines in comprehensions.
+
+    Reasoning:
+        We do this for consistency.
+
+    Solution:
+        Either place comprehension on a single line or ensure that action,
+        for loops, and condition are all on different lines.
+
+    Example::
+
+        # Correct:
+        list = [some(number) for number in numbers]
+
+        list = [
+           some(number)
+           for numbers in matrix
+           for number in numbers
+           if number > 0
+        ]
+
+        # Wrong:
+        list = [
+            some(number) for number in numbers
+            if number > 0
+        ]
+
+    .. versionadded:: 0.15.0
+
+    """
+
+    error_template = 'Found an inconsistently structured comprehension'
+    code = 361
+
+
+@final
+class AssignToSliceViolation(ASTViolation):
+    """
+    Forbid assignment to a subscript slice.
+
+    Reasoning:
+        Assingment to a slice may lead to a list changing its size
+        implicitly and strangely which makes it hard to spot bugs.
+
+    Solution:
+        Use explicit index assignment in place of slice assignment.
+
+    Why you may disable or inline-ignore this rule?
+
+    The quite common and useful example which violates this rule
+    is inplace list replacement via ``[:]`` - this helps
+    to keep the same object reference while it content could be completely
+    erased or replaced with the new one.
+
+    One more thing: slice assignment is the only way
+    for inplace array multiple replacement when you need that.
+
+    Example::
+
+        # Correct:
+        a[5] = 1
+
+        # Wrong:
+        a[1:3] = [1, 2]
+        a[slice(1)] = [1, 3]
+
+    .. versionadded:: 0.15.0
+
+    """
+
+    error_template = 'Found assignment to a subscript slice'
+    code = 362