+++ wemake_python_styleguide/visitors/ast/naming.py
@@ -357,6 +357,9 @@ class WrongVariableAssignmentVisitor(BaseNodeVisitor):
         if not node.value:
             return
 
+        if isinstance(nodes.get_context(node), ast.ClassDef):
+            return  # This is not a variable, but a class property
+
         var_values = name_nodes.get_variables_from_node(node.value)
         for var_name, var_value in itertools.zip_longest(names, var_values):
             if var_name == var_value:
@@ -371,7 +374,11 @@ class WrongVariableAssignmentVisitor(BaseNodeVisitor):
         node: AnyAssign,
         names: List[str],
     ) -> None:
-        for used_name, count in Counter(names).items():
+        used_names = filter(
+            lambda assigned_name: not access.is_unused(assigned_name),
+            names,
+        )
+        for used_name, count in Counter(used_names).items():
             if count > 1:
                 self.add_violation(
                     best_practices.ReassigningVariableToItselfViolation(