+++ wemake_python_styleguide/visitors/ast/statements.py
@@ -4,7 +4,11 @@ from typing import ClassVar, Mapping, Optional, Sequence, Set, Union
 from typing_extensions import final
 
 from wemake_python_styleguide import constants
-from wemake_python_styleguide.compat.aliases import ForNodes, FunctionNodes
+from wemake_python_styleguide.compat.aliases import (
+    ForNodes,
+    FunctionNodes,
+    TextNodes,
+)
 from wemake_python_styleguide.logic import nodes
 from wemake_python_styleguide.logic.arguments import call_args
 from wemake_python_styleguide.logic.naming import name_nodes
@@ -357,6 +361,7 @@ class PointlessStarredVisitor(BaseNodeVisitor):
         ast.List,
         ast.Set,
         ast.Tuple,
+        *TextNodes,
     )
 
     def visit_Call(self, node: ast.Call) -> None: