+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Semantic versioning in our case means:
 - Forbids using non-trivial expressions as an argument to `except`
 - Forbids using too many variables in a tuple unpacking
 - Forbids using `float("NaN")`.
+- Forbids assigning to a slice
 - Allow `__call__` method to be asynchronous
 - Allows common strings not to be counted against string constant overuse limit
 - Forbids to unpack iterable objects to lists #1259
@@ -39,6 +40,8 @@ Semantic versioning in our case means:
 - Forbids not returning anything in functions and methods starting with `get_`
 - Forbids to use empty comment
 - Forbids using bitwise operation with boolean operation
+- Forbids inconsistent structuring of multiline comprehensions
+- Forbids to use unpythonic getters and setters such as `get_attribute` or `set_attribute`
 
 ### Bugfixes
 
@@ -51,7 +54,13 @@ Semantic versioning in our case means:
 - Fixes false positives in WPS513 and WPS323
 - Fixes false positive WPS426 if `lambda` in loop uses only its arguments
 - Fixes false negative WPS421 with `pprint.pprint`
+- Fixes WPS441 triggering when reusing variable names in multiple loops
 - Fixes false positive ImplicitEnumerateViolation on range with step #1742
+- Allows to use `_` to declare several unused variables,
+  like: `x, _, _ = coordinates()`
+- Fixes variable reassignment in class context
+- Fixes that `*'abc'` was not counted as pointless star expression
+- Fixes that `-some` was counted as overused expression
 
 ### Misc